requirements.txt
python-telegram-bot>=22.5
python-dotenv>=1.2.1
google-api-python-client>=2.187.0
google-auth-httplib2>=0.3.0
google-auth-oauthlib>=1.2.3
shapely>=2.1.2
geojson>=3.2.0
